Refactor: Move types to util
