+++ autotest-plr/src/main/java/ca/bc/gov/health/qa/autotest/plr/util/TelecommunicationType.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.health.qa.autotest.plr.web.tests.types;
+package ca.bc.gov.health.qa.autotest.plr.util;
 
 public enum TelecommunicationType {
 	PHONE("T - Telephone"),
