+++ autotest-plr/src/main/java/ca/bc/gov/health/qa/autotest/plr/web/actions/UpdateFacilitySimpleActions.java
@@ -20,11 +20,11 @@ import ca.bc.gov.health.qa.autotest.plr.web.actions.model.facility.Relationship;
 import ca.bc.gov.health.qa.autotest.plr.web.actions.model.facility.Telecommunication;
 import ca.bc.gov.health.qa.autotest.plr.web.pages.plr.facility.FacilitySection;
 import ca.bc.gov.health.qa.autotest.plr.web.pages.plr.facility.UpdateFacilityPage;
-import ca.bc.gov.health.qa.autotest.plr.web.tests.types.ElectronicAddressType;
-import ca.bc.gov.health.qa.autotest.plr.web.tests.types.IdentifierTypeName;
-import ca.bc.gov.health.qa.autotest.plr.web.tests.types.RelatedProviderIdentifierType;
-import ca.bc.gov.health.qa.autotest.plr.web.tests.types.RelationshipType;
-import ca.bc.gov.health.qa.autotest.plr.web.tests.types.TelecommunicationType;
+import ca.bc.gov.health.qa.autotest.plr.util.ElectronicAddressType;
+import ca.bc.gov.health.qa.autotest.plr.util.IdentifierTypeName;
+import ca.bc.gov.health.qa.autotest.plr.util.RelatedProviderIdentifierType;
+import ca.bc.gov.health.qa.autotest.plr.util.RelationshipType;
+import ca.bc.gov.health.qa.autotest.plr.util.TelecommunicationType;
 import ca.bc.gov.health.qa.autotest.plr.web.tests.UpdateSimpleHelper;
 import ca.bc.gov.health.qa.autotest.runner.util.log.ExecutionLogManager;
 import ca.bc.gov.health.qa.autotest.runner.util.selenium.SeleniumSession;
